123일 24-12-20 isNumberKeyInput() isNumberKey()의 형제함수 한글은 IME 처리방식이라 js에서 막을수없음 그래서 값자체에서 삭제처리해줘야함

+++ 방송과사람들(방송인중개플랫폼)/public_html/component/profile/profile-section4.php
@@ -3,63 +3,47 @@
     <section id="profile04">
         <div>
             <h4>학력전공</h4>
-            <label for="">학력 · 전공</label>
-            <input type="text" name="" id="" placeholder="추가해주세요" @click="modal = true;"/>
-            <br><br>
-            <!-- 학력 전공 등록 모달 -->
-            <div class="modal fade" :class="{'in' : modal}" id="educationModal" tabindex="-1" :style="{display : modal ? 'block' : 'none'}" @click.self="modal = false;">
-                <div class="modal-dialog" role="document">
-                    <div class="modal-content">
-                        <div class="modal-header">
-                            <button type="button" class="close" data-dismiss="modal" aria-label="Close" @click="modal = false"><i class="fa-light fa-close"></i></button>
-                            <h5 class="modal-title" id="educationModalLabel">학력 · 전공</h5>
-                        </div>
-                        <div class="modal-body">
-                            <dl>
-                                <dt>학교명</dt>
-                                <dd><input type="text" id="" placeholder="학교명 입력" v-model="school.name"></dd>
-                            </dl>
-                            <dl>
-                                <dt>전공</dt>
-                                <dd><input type="text" id="" placeholder="전공 입력" v-model="school.major"></dd>
-                            </dl>
-                            <dl>
-                                <dt>상태</dt>
-                                <dd>
-                                    <select v-model="school.state">
-                                        <option value="이수">이수</option>
-                                        <option value="졸업">졸업</option>
-                                        <option value="재학">재학</option>
-                                        <option value="휴학">휴학</option>
-                                    </select>
-                                </dd>
-                            </dl>
-                            <dl>
-                                <dt>증빙자료 첨부(선택)</dt>
-                                <dd>
-                                    <div id="addFile" class="addFile">
-                                        <label class="btn" for="school_file">파일 첨부</label>
-                                        <span>{{school.upfile ? school.upfile.name : '증빙자료 파일 첨부'}}</span>
-                                        <input type="file" id="school_file" style="display: none;" @change="updateSchoolFile">
-                                    </div>
-                                </dd>
-                            </dl>
-                            <div class="box_blue" id="tip">
-                                <p>TIP</p>
-                                <ul>
-                                    <li>증빙 자료 첨부하시면 담당자 검토후 확인 마크<i class="fa-sharp fa-solid fa-badge-check"></i>를 달아드립니다.</li>
-                                    <li><strong>첨부가능자료 : 재학증명서, 졸업증명서,성적증명서</strong></li>
-                                </ul>
-                            </div>
-                        </div>
-                        <div class="modal-footer">
-                            <!--<button type="button" class="btn btn-secondary" data-dismiss="modal">닫기</button>-->
-                            <button type="button" class="btn btn-primary" @click="postSchool">적용</button>
-                        </div>
+            <dl>
+                <dt>학교명</dt>
+                <dd><input type="text" id="" placeholder="학교명 입력" v-model="school.name"></dd>
+            </dl>
+            <dl>
+                <dt>전공</dt>
+                <dd><input type="text" id="" placeholder="전공 입력" v-model="school.major"></dd>
+            </dl>
+            <dl>
+                <dt>상태</dt>
+                <dd>
+                    <select v-model="school.state">
+                        <option value="이수">이수</option>
+                        <option value="졸업">졸업</option>
+                        <option value="재학">재학</option>
+                        <option value="휴학">휴학</option>
+                    </select>
+                </dd>
+            </dl>
+            <dl>
+                <dt>증빙자료 첨부(선택)</dt>
+                <dd>
+                    <div id="addFile" class="addFile">
+                        <label class="btn" for="school_file">파일 첨부</label>
+                        <span>{{school.upfile ? school.upfile.name : '증빙자료 파일 첨부'}}</span>
+                        <input type="file" id="school_file" style="display: none;" @change="updateSchoolFile">
                     </div>
-                </div>
+                </dd>
+            </dl>
+            <div class="box_blue" id="tip">
+                <p>TIP</p>
+                <ul>
+                    <li>증빙 자료 첨부하시면 담당자 검토후 확인 마크<i class="fa-solid fa-badge-check"></i>를 달아드립니다.</li>
+                    <li><strong>첨부가능자료 : 재학증명서, 졸업증명서,성적증명서</strong></li>
+                </ul>
             </div>
-            <!-- 학력 전공 리스트 -->
+            <button class="btn btn_middle btn_blue2" @click="postSchool">등록 및 추가</button>
+            <hr>
+            <br>
+
+            <br>
             <div class="tag">
                 <template v-for="item in schools">
                     <span>{{item.name}}·{{item.major}}·{{item.state}}·
@@ -72,56 +56,39 @@
         </div>
         <div>
             <h4>보유 자격증</h4>
-            <label for="">보유 자격증</label>
-            <input type="text" name="" id="" placeholder="추가해주세요" @click="modal2 = true;"/>
-            <br><br>
-            <!-- 학력 전공 등록 모달 -->
-            <div class="modal fade" :class="{'in' : modal2}" tabindex="-1" :style="{display : modal2 ? 'block' : 'none'}" @click.self="modal2 = false;">
-                <div class="modal-dialog">
-                    <div class="modal-content">
-                        <div class="modal-header">
-                            <button type="button" class="close" data-dismiss="modal" aria-label="Close"><i class="fa-light fa-close"></i></button>
-                            <h5 class="modal-title" id="licenseModalLabel">보유 자격증</h5>
-                        </div>
-                        <div class="modal-body">
-                            <dl>
-                                <dt>자격증명</dt>
-                                <dd><input type="text" id="" placeholder="자격증명 입력" v-model="certify.name"></dd>
-                            </dl>
-                            <dl>
-                                <dt>발급기관</dt>
-                                <dd><input type="text" id="" placeholder="발급기관 입력" v-model="certify.agency"></dd>
-                            </dl>
-                            <dl>
-                                <dt>발급일</dt>
-                                <dd><input type="date" v-model="certify.issue_date"></dd>
-                            </dl>
-                            <dl>
-                                <dt>증빙자료 첨부(선택)</dt>
-                                <dd>
-                                    <div id="addFile" class="addFile">
-                                        <label class="btn" for="certify_file">파일 첨부</label>
-                                        <span>{{certify.upfile ? certify.upfile.name : '증빙자료 파일 첨부'}}</span>
-                                        <input type="file" id="certify_file" style="display: none;" @change="updateCertifyFile">
-                                    </div>
-                                </dd>
-                            </dl>
-                            <div class="box_blue" id="tip">
-                                <p>TIP</p>
-                                <ul>
-                                    <li>증빙 자료 첨부하시면 담당자 검토후 확인 마크<i class="fa-sharp fa-solid fa-badge-check"></i>를 달아드립니다.</li>
-                                    <li><strong>첨부가능자료 : 자격증사본</strong></li>
-                                </ul>
-                            </div>
-                        </div>
-                        <div class="modal-footer">
-                            <!--<button type="button" class="btn btn-secondary" data-dismiss="modal">닫기</button>-->
-                            <button type="button" class="btn btn-primary" @click="postCertify">적용</button>
-                        </div>
+            <br>
+            <dl>
+                <dt>자격증명</dt>
+                <dd><input type="text" id="" placeholder="자격증명 입력" v-model="certify.name"></dd>
+            </dl>
+            <dl>
+                <dt>발급기관</dt>
+                <dd><input type="text" id="" placeholder="발급기관 입력" v-model="certify.agency"></dd>
+            </dl>
+            <dl>
+                <dt>발급일</dt>
+                <dd><input type="date" v-model="certify.issue_date"></dd>
+            </dl>
+            <dl>
+                <dt>증빙자료 첨부(선택)</dt>
+                <dd>
+                    <div id="addFile" class="addFile">
+                        <label class="btn" for="certify_file">파일 첨부</label>
+                        <span>{{certify.upfile ? certify.upfile.name : '증빙자료 파일 첨부'}}</span>
+                        <input type="file" id="certify_file" style="display: none;" @change="updateCertifyFile">
                     </div>
-                </div>
+                </dd>
+            </dl>
+            <div class="box_blue" id="tip">
+                <p>TIP</p>
+                <ul>
+                    <li>증빙 자료 첨부하시면 담당자 검토후 확인 마크<i class="fa-sharp fa-solid fa-badge-check"></i>를 달아드립니다.</li>
+                    <li><strong>첨부가능자료 : 자격증사본</strong></li>
+                </ul>
             </div>
-            <!-- 학력 전공 리스트 -->
+            <button class="btn btn_middle btn_blue2" @click="postCertify">등록 및 추가</button>
+            <br>
+
             <div class="tag">
                 <template v-for="item in certifies">
                     <span>{{item.name}}·{{item.agency}}·{{item.issue_date}}·
@@ -146,7 +113,8 @@
             return {
                 jl : null,
                 filter : {
-                    member_idx : this.user.mb_no
+                    member_idx : this.user.mb_no,
+                    approval : '1'
                 },
                 data : {
 
@@ -189,6 +157,7 @@
         },
         methods: {
             async deleteSchool(idx) {
+                if(!confirm("삭제하시겠습니까? 해당 데이터는 삭제시 바로 삭제됩니다.")) return false;
                 var method = "delete";
                 var filter = {idx : idx }
 
@@ -206,7 +175,7 @@
                 var res = await this.jl.ajax(method,obj,"/api/member_school.php");
                 if (res) {
                     this.jl.log(res)
-                    alert("추가되었습니다.")
+                    alert("추가되었습니다. 승인 완료가 되면 노출됩니다.")
                     this.school = this.jl.initObject(this.school)
                     this.school.member_idx = this.user.mb_no;
                     this.getSchool();
@@ -222,6 +191,7 @@
                 }
             },
             async deleteCertify(idx) {
+                if(!confirm("삭제하시겠습니까? 해당 데이터는 삭제시 바로 삭제됩니다.")) return false;
                 var method = "delete";
                 var filter = {idx : idx }
 
@@ -239,7 +209,7 @@
                 var res = await this.jl.ajax(method,obj,"/api/member_certify.php");
                 if (res) {
                     this.jl.log(res)
-                    alert("추가되었습니다.")
+                    alert("추가되었습니다. 승인 완료가 되면 노출됩니다.")
                     this.certify = this.jl.initObject(this.certify)
                     this.certify.member_idx = this.user.mb_no;
                     this.getCertify();
